+++ src/api/message.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct MessagesResponse {
-    pub anchor: Option<u32>,
+    pub anchor: Option<i64>,
     pub found_newest: bool,
     pub found_oldest: bool,
     pub found_anchor: bool,
@@ -13,22 +13,22 @@ pub struct MessagesResponse {
 pub struct Message {
     pub content: String,
     pub content_type: String,
-    pub id: u32,
+    pub id: i64,
     pub sender_email: String,
     pub sender_full_name: String,
-    pub sender_id: u32,
-    pub stream_id: u32,
+    pub sender_id: i64,
+    pub stream_id: i64,
     pub subject: String,
     pub timestamp: u64,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct MessageEvent {
-    pub id: i32,
+    pub id: i64,
     pub message: Message
 }
 
 #[derive(Debug)]
 pub struct PostResponse {
-    pub id: u32
+    pub id: i64
 }
